+++ CHANGELOG.md
@@ -7,6 +7,19 @@ Version 1.X.X *(In Development)*
  * New assertions:
    * `android.app.Instrumentation.ActivityMonitor`
    * `android.app.Instrumentation.ActivityResult`
+   * `android.app.ListActivity`
+   * `android.preference.CheckBoxPreference`
+   * `android.preference.DialogPreference`
+   * `android.preference.EditTextPreference`
+   * `android.preference.ListPreference`
+   * `android.preference.MultiSelectListPreference`
+   * `android.preference.PreferenceActivity`
+   * `android.preference.Preference`
+   * `android.preference.PreferenceGroup`
+   * `android.preference.PreferenceScreen`
+   * `android.preference.RingtonePreference`
+   * `android.preference.SwitchPreference`
+   * `android.preference.TwoStatePreference`
    * `android.util.AttributeSet`
    * `android.util.Pair`
  * Added convenience methods for orientations to `Display`, `LinearLayout`, and